+++ utils/utils.go
@@ -1,0 +1,11 @@
+package utils
+
+import (
+	"log"
+)
+
+func CheckError(err error) {
+	if err != nil {
+		log.Println(err)
+	}
+}
